+++ src/main/java/net/grid/vampiresdelight/common/item/HunterConsumableItem.java
@@ -3,7 +3,8 @@ package net.grid.vampiresdelight.common.item;
 import de.teamlapen.vampirism.VampirismMod;
 import de.teamlapen.vampirism.api.VReference;
 import de.teamlapen.vampirism.core.ModEffects;
-import net.grid.vampiresdelight.common.VDFoodValues;
+import net.grid.vampiresdelight.common.food.VDFoodValues;
+import net.grid.vampiresdelight.common.food.FoodFeatures;
 import net.grid.vampiresdelight.common.utility.*;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.food.FoodProperties;
@@ -27,6 +28,7 @@ import java.util.List;
 
 public class HunterConsumableItem extends Item {
     private final FoodProperties hunterFood;
+    private final FoodFeatures foodFeatures;
     private final boolean hasFoodEffectTooltip;
     private final boolean hasCustomTooltip;
     private final boolean containsGarlic;
@@ -34,6 +36,7 @@ public class HunterConsumableItem extends Item {
     public HunterConsumableItem(Properties properties) {
         super(properties);
         this.hunterFood = null;
+        this.foodFeatures = null;
         this.hasFoodEffectTooltip = false;
         this.hasCustomTooltip = false;
         this.containsGarlic = true;
@@ -42,6 +45,16 @@ public class HunterConsumableItem extends Item {
     public HunterConsumableItem(Properties properties, boolean hasFoodEffectTooltip) {
         super(properties);
         this.hunterFood = null;
+        this.foodFeatures = null;
+        this.hasFoodEffectTooltip = hasFoodEffectTooltip;
+        this.hasCustomTooltip = false;
+        this.containsGarlic = true;
+    }
+
+    public HunterConsumableItem(Properties properties, FoodFeatures foodFeatures, boolean hasFoodEffectTooltip) {
+        super(properties);
+        this.hunterFood = null;
+        this.foodFeatures = foodFeatures;
         this.hasFoodEffectTooltip = hasFoodEffectTooltip;
         this.hasCustomTooltip = false;
         this.containsGarlic = true;
@@ -50,6 +63,16 @@ public class HunterConsumableItem extends Item {
     public HunterConsumableItem(Properties properties, FoodProperties hunterFood, boolean hasFoodEffectTooltip, boolean hasCustomTooltip, boolean containsGarlic) {
         super(properties);
         this.hunterFood = hunterFood;
+        this.foodFeatures = null;
+        this.hasFoodEffectTooltip = hasFoodEffectTooltip;
+        this.hasCustomTooltip = hasCustomTooltip;
+        this.containsGarlic = containsGarlic;
+    }
+
+    public HunterConsumableItem(Properties properties, FoodProperties hunterFood, FoodFeatures foodFeatures, boolean hasFoodEffectTooltip, boolean hasCustomTooltip, boolean containsGarlic) {
+        super(properties);
+        this.hunterFood = hunterFood;
+        this.foodFeatures = foodFeatures;
         this.hasFoodEffectTooltip = hasFoodEffectTooltip;
         this.hasCustomTooltip = hasCustomTooltip;
         this.containsGarlic = containsGarlic;
@@ -58,8 +81,13 @@ public class HunterConsumableItem extends Item {
     @Override
     public ItemStack finishUsingItem(ItemStack stack, Level level, LivingEntity consumer) {
         if (!level.isClientSide) {
-            if (containsGarlic)
+            if (containsGarlic) {
                 VDEntityUtils.cureEffect(ModEffects.SANGUINARE.get(), consumer);
+            }
+
+            if (foodFeatures != null) {
+                foodFeatures.execute(consumer);
+            }
 
             this.affectConsumer(stack, level, consumer);
         }
@@ -100,6 +128,11 @@ public class HunterConsumableItem extends Item {
         }
 
         return VDHelper.isHunter(entity) && hunterFood != null ? hunterFood : (VDHelper.isVampire(entity) ? VDFoodValues.NONE : super.getFoodProperties(stack, entity));
+    }
+
+    @Override
+    public int getUseDuration(ItemStack pStack) {
+        return (foodFeatures.getUseDuration() > 0) ? foodFeatures.getUseDuration() : super.getUseDuration(pStack);
     }
 
     /**