+++ src/main/java/net/grid/vampiresdelight/common/item/WerewolfConsumableItem.java
@@ -1,6 +1,7 @@
 package net.grid.vampiresdelight.common.item;
 
 import de.teamlapen.vampirism.VampirismMod;
+import net.grid.vampiresdelight.common.food.FoodFeatures;
 import net.grid.vampiresdelight.common.utility.VDIntegrationUtils;
 import net.grid.vampiresdelight.common.utility.VDTextUtils;
 import net.grid.vampiresdelight.common.utility.VDTooltipUtils;
@@ -26,6 +27,7 @@ import java.util.List;
 
 public class WerewolfConsumableItem extends Item {
     private final FoodProperties werewolfFood;
+    private final FoodFeatures foodFeatures;
     private final boolean hasFoodEffectTooltip;
     private final boolean hasHumanFoodEffectTooltip;
     private final boolean hasCustomTooltip;
@@ -34,6 +36,17 @@ public class WerewolfConsumableItem extends Item {
     public WerewolfConsumableItem(Properties pProperties, FoodProperties werewolfFood) {
         super(pProperties);
         this.werewolfFood = werewolfFood;
+        this.foodFeatures = null;
+        this.hasFoodEffectTooltip = true;
+        this.hasCustomTooltip = false;
+        this.hasHumanFoodEffectTooltip = false;
+        this.hasFactionTooltip = true;
+    }
+
+    public WerewolfConsumableItem(Properties pProperties, FoodProperties werewolfFood, FoodFeatures foodFeatures) {
+        super(pProperties);
+        this.werewolfFood = werewolfFood;
+        this.foodFeatures = foodFeatures;
         this.hasFoodEffectTooltip = true;
         this.hasCustomTooltip = false;
         this.hasHumanFoodEffectTooltip = false;
@@ -43,6 +56,7 @@ public class WerewolfConsumableItem extends Item {
     public WerewolfConsumableItem(Properties pProperties, FoodProperties werewolfFood, boolean hasFoodEffectTooltip, boolean hasCustomTooltip, boolean hasHumanFoodEffectTooltip, boolean hasFactionTooltip) {
         super(pProperties);
         this.werewolfFood = werewolfFood;
+        this.foodFeatures = null;
         this.hasFoodEffectTooltip = hasFoodEffectTooltip;
         this.hasCustomTooltip = hasCustomTooltip;
         this.hasHumanFoodEffectTooltip = hasHumanFoodEffectTooltip;
@@ -52,6 +66,10 @@ public class WerewolfConsumableItem extends Item {
     @Override
     public ItemStack finishUsingItem(ItemStack stack, Level level, LivingEntity consumer) {
         if (!level.isClientSide) {
+            if (foodFeatures != null) {
+                foodFeatures.execute(consumer);
+            }
+
             this.affectConsumer(stack, level, consumer);
         }
 
@@ -91,6 +109,11 @@ public class WerewolfConsumableItem extends Item {
         }
 
         return VDIntegrationUtils.isWerewolf(entity) ? werewolfFood : super.getFoodProperties(stack, entity);
+    }
+
+    @Override
+    public int getUseDuration(ItemStack pStack) {
+        return (foodFeatures.getUseDuration() > 0) ? foodFeatures.getUseDuration() : super.getUseDuration(pStack);
     }
 
     /**